Renamed package `ui` to `gui`
