+++ cmd/gui.go
@@ -1,7 +1,7 @@
 package cmd
 
 import (
-	"github.com/habedi/gogg/ui" // update the import path as needed
+	"github.com/habedi/gogg/gui"
 	"github.com/spf13/cobra"
 )
 
@@ -10,7 +10,7 @@ func guiCmd() *cobra.Command {
 		Use:   "gui",
 		Short: "Start the Gogg GUI",
 		Run: func(cmd *cobra.Command, args []string) {
-			ui.Run()
+			gui.Run()
 		},
 	}
 	return cmd
