+++ gui/catalogue_test.go
@@ -1,11 +1,11 @@
-package ui_test
+package gui_test
 
 import (
 	"testing"
 
 	"fyne.io/fyne/v2/test"
 	"github.com/habedi/gogg/db"
-	"github.com/habedi/gogg/ui"
+	"github.com/habedi/gogg/gui"
 )
 
 // initDBForTest initializes the database for UI tests.
@@ -35,7 +35,7 @@ func TestCatalogueListUI(t *testing.T) {
 	t.Cleanup(w.Close)
 
 	// Call CatalogueListUI and verify it returns a non-nil CanvasObject.
-	co := ui.CatalogueListUI(w)
+	co := gui.CatalogueListUI(w)
 	if co == nil {
 		t.Error("CatalogueListUI returned nil, expected a valid fyne.CanvasObject")
 	}
@@ -51,13 +51,13 @@ func TestSearchCatalogueUI(t *testing.T) {
 	t.Cleanup(w.Close)
 
 	// Test search by title.
-	co := ui.SearchCatalogueUI(w, "Test", false)
+	co := gui.SearchCatalogueUI(w, "Test", false)
 	if co == nil {
 		t.Error("SearchCatalogueUI (by title) returned nil, expected a valid fyne.CanvasObject")
 	}
 
 	// Test search by ID (assuming "1" is valid).
-	co2 := ui.SearchCatalogueUI(w, "1", true)
+	co2 := gui.SearchCatalogueUI(w, "1", true)
 	if co2 == nil {
 		t.Error("SearchCatalogueUI (by ID) returned nil, expected a valid fyne.CanvasObject")
 	}
@@ -74,7 +74,7 @@ func TestRefreshCatalogueUI(t *testing.T) {
 
 	// Call RefreshCatalogueUI; we simply check that it runs without panic.
 	// (This function opens a dialog and runs asynchronously.)
-	ui.RefreshCatalogueUI(w)
+	gui.RefreshCatalogueUI(w)
 	// In a real test you might wait for completion or verify side effects.
 }
 
@@ -90,6 +90,6 @@ func TestExportCatalogueUI(t *testing.T) {
 	// Call ExportCatalogueUI for both supported formats.
 	// Since the export function opens a file save dialog,
 	// we only verify that the call does not panic.
-	ui.ExportCatalogueUI(w, "csv")
-	ui.ExportCatalogueUI(w, "json")
+	gui.ExportCatalogueUI(w, "csv")
+	gui.ExportCatalogueUI(w, "json")
 }
