+++ gui/download.go
@@ -1,5 +1,5 @@
 // ui/download.go
-package ui
+package gui
 
 import (
 	"fmt"
